requirements.txt
gradio
openai-whisper
translate
python-dotenv